==> Analysis/01_Data_Validation.ipynb <==
# **Global E-Commerce Transactions Analysis Using Python**

# 1. Data Validation:

This project analyzes 201,110 global e-commerce transactions covering both Indian (national) and international markets.

Key Goals:
- Understand transaction trends.
- Validate data quality.
- Prepare dataset for further analysis.

---

# Libraries:

import pandas  as pd 
import numpy   as np
import matplotlib.pyplot as plt
import seaborn as sns

pd.set_option('display.max_columns', None)
pd.set_option('display.float_format', '{:.2f}'.format)

# Dataset:

df = pd.read_csv('Global_E-Commerce_Transactions.csv')
df.head()

---
Data Overview:

1. Number of Rows & Columns:


df.shape

---
2. Column Names:

df.columns

---
3. Data Type:

df.info()

---
4. Summary Statistics:

df.describe(include = "all").T

---
5. Missing Values Check:

df.isnull().sum()

---
6. Duplicate Check:

df.duplicated().sum()

---

# Observations:

- Dataset contains **201,110 rows & 15 columns**.
- Covers multiple **countries and cities**.
- No structural issues found at import stage.
- Ready for cleaning and preprocessing.

---
